Added config group and default for optims, moved into config python script
